Remove Nicotine Moodlet Spam (#521)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Comments out the nicotine moodlet because it's broken and incomplete.
For as long as this "feature" has existed, I have complained that the
popup text of "I feel as if I have been standing my entire life and I
just sat down." is _faaaar_ too frequent, often to the point where it
visually overwhelms me. It is essentially spam and I want it gone. Maybe
I didn't remove it the correct way, but I just need this out of my line
of sight.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] ur momma

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- remove: no more moodlet text spam when you smoke cigarettes finally
thank fuck jesus christ