Tastes Like.... Green. (#522)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

PREFACE: this is entirely a spite PR, though I wish no ill will or
malice towards the original writer of these flavor profiles. Once upon a
time, I was helping to come up with flavor descriptions for unique
drinks in DeltaV. I suggested "tastes like green" for absinthe because
it's a very clear Star Trek reference and Star Trek is one of the
inspirations for this game and its predecessor and I also just really
like Star Trek. Next thing I know, the drink is flavored like "green
death". So I'm changing it to "tastes like... green" to Restore My
Original Vision For Space Station 14 As A Whole.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Slightly modified absinthe flavor. It tastes like... *sniff*...
hmm... like green.