Crusher Axe and Glaive Tweaks (#524)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Tweaked the Crusher Axe to have its lifesteal back, its increased damage
on wield and removed the wield requirements for swinging, and added the
old attack speed values for both axe and glaive. Also removes the
pressure requirements for both of them.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Tweaked the crusher axe to have its lifesteal and increased
damage on wield back.
- tweak: Tweaked both the crusher axe and glaive to its old speed
values, and removed the pressure and wield requirements.