More Loadout Corrections (#503)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Deduplicates loadout options, (with preference to base EE over floof)
and removes some weird loadout grouos

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Delete Duplicate Loadout Items
- [X] Remove Unecessary Eyes Loadout Group
- [x] Remove Unecessary Fun Loadout Group
- [X] Remove Unecessary Glove Loadout Group
- [X] Remove Unecessary Head Loadout Group
- [x] Remove Unecessary Mask Loadout Group
- [x] Remove Unecessary Neck Loadout Group
- [x] Remove Unecessary Outer Loadout Group
- [x] Remove Unecessary Shoe Loadout Group
- [x] Remove Unecessary Uniform Loadout Group
- [x] Move Items Into Correct Category/Jobs

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: MORE loadout changes, be sure to make sure your loadouts are
updated