Detail Examine Range Constraint Removal (#530)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

The variable which determines if you're in close enough proximity to an
entity to see its detailed examine text is now always true. Removing it
completely and setting the disabled flag to be false caused some issues,
so I opted to do it this way instead. (I have no idea what I'm doing.)

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Removed the range limitation on detail examine.

Co-authored-by: MajorMoth <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>