IPC "Visor Eyes" Marking (#534)
# Description

Once upon a time there was a SS13 fork that had IPC markings that were
cool and unique to that fork.

I hit up the guy who added these markings and said, "Hey, can I port
these? I mea I know it's AGPL anyway". They responded, "I dunno man...
they're kind of dog ass...". Being the astute subtext reader I am I took
this as a resounding "yes"

As an additional little treat I also added a marking to cover the screen
because unlike that server, adding a face marking does not remove the
black screen part of the head. The screen is designed for the default
IPC head and may not work as well for others.

It's recolorable

---

# TODO
- [x] Port IPC screen visor sprites
- [x] Add IPC screen visor marking
- [x] (Optional) add "screen covering"

---

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/a67ede62-0112-4884-9907-956b388b7361)


![image](https://github.com/user-attachments/assets/16ccd8dc-3f12-4b80-b28e-af57650cd4d8)

</p>
</details>

---

# Changelog

:cl:
- add: IPCs have a visor-like face option and a "screen covering" to
hide their face.