Asterisk/Shoko Fixes (#513)
Added Plasma to Asterisk's Chem Lab at Round Start Re-mapped the Mining
Shuttle Dock (still was deleted from the Migration issue and saved, ergo
no shuttle dock)

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Added ten Plasma to Asterisk's Chem lab.
Placed a missing tile under a door at Asterisk's new dorms.
Fixed a typo on Dagmire's description.

Re-mapped the mining shuttle dock for Shoko.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Gave roundstart Asterisk Chemistry Plasma
- fix: Fixed Shoko's Mining Shuttle Dock spawn