Security Sawn Off (#532)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds a sawn off loaded with non-lethals and a beanbag magazine to the
security loadout. Brings back the shell box sprites from Frontier after
they got reverted by a past merge.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added a sawn off shotgun loaded with non-lethals and a beanbag
magazine to the security loadouts. Please practice gun safety and
self-restraint.
- tweak: Tweaked shotgun shell boxes sprites to the ones from Frontier
once again.