Engineering Stuff (#533)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Ports glow in the dark hi-vis jackets as well as the magmantle from EE.
Ports the magboots refactor PR as well.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added the magmantle, a cloak that functions as slower magboots.
Available in the species loadout for species that lack feet slots in
certain departments or jobs.
- tweak: Tweaked Hi-Vis jackets to glow in the dark.

---------

Co-authored-by: Will-Oliver-Br <164823659+Will-Oliver-Br@users.noreply.github.com>
Co-authored-by: kosticia <kosticia46@gmail.com>
Co-authored-by: Pon <pongozdalski@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>