Ports More Hyperstation IPC Markings (#542)
# Description

Turns out there was a bit more Hyperstation IPC content than I thought,
and I figured it'd be a good idea to port it here.

This PR includes:

- "Sleek" and "Jointed" part sets
- Rounded head
- Five new antennae (iDog ears up/down, headphones, cat ears, fins)

It also moves the IPC visor screens I added last PR to a proper "IPC"
folder (sorry) and moves that down by a pixel because as it turns out
our IPC sprites are shorter than the server the assets are from.

---

# TODO

- [x] Port antennae
- [x] Port limbs + head
- [ ] Port animated screens?

---

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/6e14eff3-9d13-44e0-976d-627a08a076f2)

![image](https://github.com/user-attachments/assets/e4f61bb9-f369-4709-b264-b84273cce306)

![image](https://github.com/user-attachments/assets/b24cc38d-98f9-4644-81bc-c1f8040eb308)

![image](https://github.com/user-attachments/assets/254a339c-848e-493a-892d-4388cb9cb319)

![image](https://github.com/user-attachments/assets/fc15c401-c7ba-4dce-9e7f-6eb0e8ec3cbe)

![image](https://github.com/user-attachments/assets/56f84139-7e09-4756-91a4-6ec6fed314cd)

</p>
</details>

---

# Changelog

:cl:
- add: Added new antennae and cosmetic limbs for IPCs.
- fix: Moved IPC visor faces down by one pixel.