Introduction to Stinky Steve's Space Weed Emporium (#548)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Added a new pack of 'cigarettes', named "Stinky Steve's packet". It
contains 10 common marijuana joints, and can be obtained via hacking the
Shadycigs vendor or by picking it out in the Loadout menu for 1 point.

Codewise, this adds the packs.yml as well as related paths to the Den
repo, allowing for more smoke-related assets in the future. The only
non-den area touched was cigs.yml , but I couldn't find a way to assign
the pack to the machine without modifying the inventory itself. I
commented it with a label, so I don't think anyone will be FURIOUS with
me.

---


# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Stinky Steve's Weed Emporium

Co-authored-by: Savante-Sam <vinnyreed2@gmail.com>