Made Wall Lockers Airtight (#553)
# Description

When you get into a wall locker, it would use the outside atmosphere.
Now, it is "airtight". I assume that means you can still deplete the
oxygen or whatever, but at least you don't die super quick.

Ideally, it would grab the atmosphere at the entry position whenever
it's opened and copy that for inside, i feel like this is good enough
for now?

# Changelog
 
 - Made wall lockers airtight

Co-authored-by: Cam <Nop>