Add Aprons to Loadouts, New Waist Aprons (#559)
# Description

This PR adds the following:

1. Bartender apron in bartender loadout
2. Chef apron in outer loadout + recolorable (because the chef apron is
white)
3. Adds four new "waist aprons": plain, checkered, horizontal and
vertical stripes. These can be worn either on the belt or in the outer
layer, but they are in the "belts" loadout category.

I added the waist aprons to the uniform printer as well, but they're a
little gormless without full RGB editing so I suggest using spray paint
on them or something. Waist aprons have a few storage slots, much like
other "outerwear" items with storage. Also added the plain one to chef
uniform vendor as it fits nicely with the other white apron.

PR is mostly made for character customization and funsies reasons. It's
time to dine

---

<details><summary><h1>Media</h1></summary>
<p>


![2025-05-18_21-08](https://github.com/user-attachments/assets/e41ee9a3-1db4-4cd1-b98b-0f6a85caa01b)

![2025-05-18_21-09](https://github.com/user-attachments/assets/4c8f4a51-ce4d-412c-94cd-04ad7fb3ffda)

![2025-05-19_00-34](https://github.com/user-attachments/assets/967d7c5f-41f3-43d6-bd25-ddd8f9e1ee68)

![2025-05-19_00-35](https://github.com/user-attachments/assets/e918a330-f5a8-46b7-9fbc-a90a8b939abd)

![2025-05-19_00-35_1](https://github.com/user-attachments/assets/7206bc02-2ac5-4883-8ca7-8ef057d6eaba)

![2025-05-19_00-35_2](https://github.com/user-attachments/assets/3992bcc5-fff7-46e9-bd12-64f7782e76be)

![2025-05-19_00-35_3](https://github.com/user-attachments/assets/79fcea0d-43c3-48de-89bb-bea92c91c91b)

</p>
</details>

---

# Changelog

:cl:
- add: Added bartender apron to the bartender loadout, and a recolorable
white apron to the civilian "Outer" loadout.
- add: Added a new, fashionable "waist apron". Add it in belt loadouts
or make it at a uniform printer. Comes in four different patterns.