Egg Layer Trait + Other Trait Related Changes (#562)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds the Egg Layer trait for 1 cost, removes the restriction on the
Honey Stomach trait but increases cost to 1. Ovinia language now able to
be taken as trait

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added Egg Layer trait for 1 cost, now you can make egg!
- tweak: Removed species restriction on Honey Stomach trait, made it
cost 1 to match egg layer.
- add: Added the Ovinia natural language to the trait list for the
typical cost