Disable Egg Layer Trait (Pending Changes/Removal) (#568)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Moving the egg layer trait more towards being a kink trait, making the
eggs that it creates inert, also room to make variant eggs for more
species specific eggs (spriting is hard so i aint making them)

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] New base egg used for trait (not for cooking)
- [ ] Descriptions and name for base egg for clarity to other players
what it is, so it is not mistaken

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Egg Layer trait made inert and not usable for cooking