Corgium Tweaks (#541)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Should make the corgium polymorph revert on death/crit, as well as
revert after 30 seconds on polymorph.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Tweaked Corgium to be less permanent, reverting on crit/death,
as well as reverting after 30 seconds.

---------

Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>