Rework Nekomimetic Language (#535)
# Description

We were discussing how the Nekomimetic language in its current state is
kind of off-putting and bad-feeling, effectively just a random mash of
"moe" buzzwords. The in-game description even describes it as "sounding
like broken Japanese". The vibes are kinda bad and I think there's ways
to make a memey jokey Catgirl Language that has less bad vibes.

As a note, this PR also removes Nekomimetic from Oni and gives them Sol
Common. I don't think it makes much sense for Oni to have Nekomimetic
after this rework, and it didn't make much sense for them to have it in
the first place. I would like for Oni to have their own language, but
I've deemed it out of scope for this PR, and it would probably come down
to the loremasters on Discord to come up with ideas for that.

I would like to hear feedback from people on this PR for how these
changes can be improved.

---

# TODO
- [x] Change Nekomimetic's syllable list to be not broken Japanese.
- [x] Fix references to Nekomimetic language in-game to reflect this
change.
- [x] Figure out what to do about Oni.

---

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/68b4a679-5664-4eeb-860c-f85bde6800fa)

![image](https://github.com/user-attachments/assets/0c001625-2a56-4232-b8ce-24908b686840)

</p>
</details>

---

# Changelog

:cl:
- tweak: Nekomimetic, the language used by Felinids, has a new font and
completely changed vocabulary.
- remove: Oni lack a racial language now (because it makes little sense
for them to be meowing like cats). This change is intended to be
temporary.

---------

Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>