Port Juggernaut Sprites From Delta V (#577)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Ports PR #3364 from Delta-V, respriting the juggernaut helmet to be more
sensible looking, including variants for humans, vox, harpies, vulps,
and feroxis. As well as resolving some issues here and there involving
their sprites.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Ported a juggernaut suit resprite from Delta-V that changes the
helmets to be less of a dome and have more species variants.

---------

Co-authored-by: boogiebogus <53976699+boogiebogus@users.noreply.github.com>