Prescription Hud and Handheld Radio Tweaks (#589)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Removes the glue from the prescription hud recipes, adds the handheld
radio to the protolathe and cargo techfab recipe lists and the security
handheld to the security techfab. Updates the security handheld radio
icons to be consistent with the normal handhelds.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Removed glue from the prescription hud recipes.
- tweak: Tweaked handheld radios (both basic and security ones) to be
craftable in their respective techfabs.
- tweak: Updated the security handheld radio icons to be consistent with
the basic handhelds.