Shadowkin Psionics Changes (#593)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Title-
Shadowkin cant just teleport away on den so.. not necessary
As for traits, they should've been able to take them from the beginning

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Shadowkin no longer have to be blackeye to be a perma prisoner.
- tweak: Shadowkin can now select psionic traits if they dont have black
eyes.