Nerf Healing Word & Tweak Breath of Life (#594)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Nerfs healing word closer to that of Tricordrazine, though this is still
stronger.
Nerfs include
- Removal of Caustic Healing
- Removal of Asphyxiation Healing
- Removal of Blood loss Healing
- Removal of Genetic Healing
- Removal of Rot Reduction
- Removal of ability to use it on yourself
- No longer works on corpses

Largely this is to give medical more opportunity to actually do their
job and not have a psion walk in and heal everyone of all their
ailments. As well as giving Regeneration it's spot back for self
healing. And Revivify the only power to reduce rot (as it is intended
for use on corpses)

To follow along Breath of Life no longer works on the living.

The goal is partially to bring Healing Word back into being a trait
ability, but at least with this, chaplains or people who random roll it
wont be Medical but better

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Healing word nerfed to a level similar, but stronger than
tricordrazine.
- tweak: Healing word no longer works on corpses.
- tweak: Breath of Life no longer works on the living.