Injectables Tweaks (#596)
# Description

Many tweaks to injectables, see changelog. On the technical side
(doesn't need to be in the changelog):
I moved all prefilled syringes and their parent to their own
syringes.yml file.
Also moved the advanced combat medipen to the file with all the other
medipens.

---

# Changelog

:cl:
- tweak: Disposable syringes now inject faster.
- tweak: All prefilled syringes (medkits, etc.) are now the disposable
variety.
- tweak: Tweaked the dosages of certain prefilled syringes.
- fix: Every autoinjector now injects the entirety of its contents upon
use. This is how it was before funkychem.
- fix: Standardized the naming convention of prefilled syringes. Should
be consistent across the board now.