Traitor Telecrystal Tweaks (#605)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adjusts the thief telecrystal to the appropriate converted values from
10 -> 50, increases the traitor telecrystal count from 100 -> 115 to
hopefully give them a slight edge.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Tweaked the thief toolbox telecrystal from 10 -> 50 as it
should have been with the TC rework.
- tweak: Tweaked the traitor telecrystal from 100 -> 115 to give them a
slight edge.