Localization Patch + 4 Compatibility Resprite (#602)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

- Fixed Localization file for Resomi's markings, hopefully this is the
last time I'm gonna deal with this
- Added sprites for duravest, plate carrier, riot suit and cargo winter
coat due to them being essentials. (Bug report)

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [X] Localization file patch for resomi
- [x] Resprite for resomi-fitting outer armors from DeltaV.

---


# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: RSI for Resomi version of DeltaV's duravest, riot suit,
platecarrier and CD's default cargo winter jacket.
- tweak: Resomi markings' localization files

---------

Co-authored-by: chly <haychiethewonderfulsnivy@gmail.com>