Salvage Weapon Tweaks + Addition (#613)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

A couple stat tweaks for certain kinetic weapons as well as a new katana
sheath available in the salvage specialist loadouts.

Stat Tweaks are as follows:
- Crusher Axe and Glaive attack speed adjusted from 0.65 to 0.8 seconds
per swing.
- Crusher Axe price in the salvage points shop increased to 2000 due to
being able to function in pressured environments.
- Kinetic Spear attack speed increased from 1.5 to 0.8 seconds per
swing, along with an increase in melee range.
- Kinetic Spear's slash damage increased from 5 to 10, for a total of 10
blunt and 10 slash.
- Kinetic Spear gains a leech on hitting marked targets for 10 Brute and
5 Burn healing.
- Kinetic Hammer's attack speed increased from 1.5 to 1 second per
swing.
- Kinetic Hammer's blunt damage increased from 17.5 to 25.
- Kinetic Machete's shield cooldown boost on hitting marked targets
increased from 20 seconds to 40 seconds of cooldown boost.


---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added a katana sheath fit for a single katana, available in the
salvage specialist loadout for veterans in the workplace.
- tweak: Tweaked some stats for the crushers and kinetic weapons with
overall improvements, view PR for details.