Paradox Anomaly Rework (#646)
:cl:
- tweak: Paradox Anomaly has been given the Thief treatment and is now
an RP free agent role without any kill objectives. Your goal is to work
with your clone to advance round roleplay or explore your new self
alone. It is highly encouraged to use OOC discussions to do this.
- tweak: Paradox Anomalies now have a little examine message that will
tell you they are not the real person. This is to prevent any
accidentally uncomfortable situations, and you should use this
information to prevent that which you are not comfortable with while not
overly metagaming. Correct usage of this is to inform the paradox when
you are uncomfortable in LOOC, while acting normal, if a bit suspicious
depending on knowledge of the person, in round.
- fix: A Paradox Anomaly will never spawn without candidates.