.50 Shell Renamed to 12 Gauge (#651)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

you dont describe (90% of) shotgun shellswith a bore value, and these
shells are aleady described as 12 gauge in game, shouldnt effect any
players from knowing what these are for

also renames shotgun shells incendiary to dragons breath, a still self
describing and a little more accurate to what it should be called

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: .50 Shell renamed to 12 gauge for better clarity
- tweak: Incendiary shells are now dragons breath for a little bit more
accuracy (insert john wick reference here)