Cherrypick DeltaV Holofans (#620)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Cherry-picks the holofans from delta PR
[#3462](https://github.com/DeltaV-Station/Delta-v/pull/3462) which makes
holofans use charges rather than battery.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Figure out how to make the sounds only activate on placement,
rather than making the holofan projector hum

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/bcde097f-d091-4128-8f3e-e0c0c7cdbdff)

![image](https://github.com/user-attachments/assets/99fd4ef3-cbea-4395-af34-e6b196f22c6f)

![image](https://github.com/user-attachments/assets/6efe9003-2714-4194-b461-67a56d7bf83a)

![image](https://github.com/user-attachments/assets/17f2b979-b1d0-41d8-896f-371dbe341ded)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: cherrypicked holofans from delta

---------

Co-authored-by: deltanedas <39013340+deltanedas@users.noreply.github.com>
Co-authored-by: Cami <147159915+Camdot@users.noreply.github.com>