Cherrypick #3875 From DeltaV (IPC Improvements + Synthetic Lizard) (#656)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description
<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds snout markings, as well as a component which allows for using
Vulpkanin headpieces on IPCs that have snouts equipped.
It also allows for different (non body) colored hair. Also ports
Synthetic Lizards from Iris (SS13) and Zenghu Pharmaceuticals Alternate
from MonkeStation (SS13).

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Look into making wag component hide action if tail not equipped

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


https://github.com/user-attachments/assets/6d7f263c-7724-472c-8fc0-3b76fdc456fb

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: synthliz from Iris and Alt Zenghu from Monkestation (SS13 Ports)
- add: IPCs with snouts will get vulpkanin helmets.
- tweak: enabled wagging tails on IPC
- tweak: whitelist snouts from vulps, rodentia, and feroxi