Hardsuit Syntax Fixes + ERT Chap & BSO Hardsuit Names (#671)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Helmets modified to match and be more clear in their descriptions
BSO renamed to Aegis
Chaplain renamed to Elymas

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Hardsuit helmet names now fit their hardsuits
- tweak: BSO tacsuit renamed to "Aegis" tacsuit
- tweak: Chaplain ERT Hardsuit renamed to "Elymas" hardsuit
- tweak: "Xíngtiān" tacsuit fits the formatting of other hardsuits now