Adds Aphrodisiac Pill Bottles to Loadout (#673)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds pill bottles of the aphrodisiac chems for a price in the lewd tab
of the loadouts. Also fixed the icons for the psionic registry.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added aphrodisiac pill bottles in the loadouts for a price, use
responsibly.