Small Drozd, C20r and FPA-90 Tweaks (#682)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Lowers the Drozd's rate of fire to match the C20r as it is frankly
ridiculous, very slightly ups the C20r's accuracy to make it the
superior choice of the two and makes the FPA-90 spawn loaded with one
mag

Drozd  -  720RPM > 600RPM
C20r  -  maximum spread angle while wielded 16 > 13
FPA-90 uplink bundle  -  +1 mag loaded


---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- Done

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Drozd's fire rate
- tweak: C20r's accuracy (slightly)
- tweak: FPA-90 has a mag when bought now