Resomi Paramedic Hardsuit Resprite (#667)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Replaced the paramedic hardsuit sprite for resomi with one that actually
matches the hardsuit that its used for.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Screenshot 2025-06-05
133946](https://github.com/user-attachments/assets/81fe0735-dfb6-4133-a30a-9c710a3f9423)
![Screenshot 2025-06-05
133950](https://github.com/user-attachments/assets/953e07ba-2a33-4ebc-9f80-fd79a1f2e7bb)
![Screenshot 2025-06-05
133954](https://github.com/user-attachments/assets/8b512879-e016-4746-b504-39ddba715508)
![Screenshot 2025-06-05
133958](https://github.com/user-attachments/assets/7ad792da-d80b-49cb-8cbd-5b4c5d93039d)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Replaced resomi paramed hardsuit sprite