Listening Outpost Camera Monitor Fixes (#683)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Listening Outpost Operatives rejoice! The Syndicate Engineers have heard
your cries, and have been hard at work on fixing the Listening Outpost
camera monitors! You may now stalk the station inhabitants and provide
proper information to your on-field Agents once again!

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- Done

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![LPO](https://github.com/user-attachments/assets/3fd35b80-2eea-462d-8dc6-48d98c725aab)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: Camera Monitor and Wireless Camera Monitor on the Listening
Outpost