Intellicard Spawn Fix and Minor Baikal Fixes (#672)
Yo yo yo! It's your Numbah One weasel with an easel! Painting sad clowns
and hanging them in my home so people think I'm cultured.
Anyways. Enough whimsy. Straight to business. Like a mullet, we keep the
professional side up front, the party in the back.

![GrpLM2KXIAAvSSB](https://github.com/user-attachments/assets/aa980713-3e2c-419e-9e24-db82bf6617dc)

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Changed the RD Locker without the Hardsuit to now have the Intellicard
in it's Fill section, where as earlier it did not.

Rotated a Bridge APC on Baikal, Replaced a Kitchen locked door that had
Passengers spawns on the other side of it with a Public Access one and
also added a single LV wire in Engineering to fix a wiring issue.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Arena Overhaul :jor:
- [x] Minor Baikal Fixes and Intellicard Spawns


---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added Intellicards to RD Lockers without Hardsuits
- fix: Fixed minor things on Baikal