Hardsuit Tweaks (#681)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Tweaked movement speed values of Security tacsuits + Paramed hardsuit
and adjusted some of the Antag hardsuits

**Station:**
--Buffs--
_Baghatur MK.II (Standard)_  -  movement speed penalty 25% > 20%
_Sulde MK.II (Riot)_                -  movement speed penalty 40% > 30%

_Sukunabikona (Paramed hardsuit)_  -  movement speed penalty 10% > 5%

**Syndicate:**
--Nerfs--
_Tianming (Commander)_  -  movement speed penalty 0% > 15%
                                           - cost 55TC > 60TC
_Guan Yu (Juggernaut)_ - walking speed penalty 10% | running speed
penalty 35% > movement speed penalty 35%
                                           -  cost 60TC > 85TC
- Radiation resistance 80% > 70%
_Shiwei (Elite)_                   -  movement speed penalty 0% > 10%
--Buffs--
_Egui (Stealth)_                  -  movement speed penalty 10% > 5%
_Xingtian (Dreadnaught)_ - walking speed penalty 30% | running speed
penalty 50% > movement speed penalty 40%

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->
Please read above in the **Description** for better details
:cl:
- tweak: Baghatur MK.II movement speed penalty decreased
- tweak: Sulde MK.II movement speed penalty decreased
- tweak: Sukunabikuna movement speed penalty decreased
- tweak: Tianming movement speed penalty increased and cost increased
- tweak: Guan Yu movement speed adjusted, cost increased and radiation
armor value decreased
- tweak: Shiwei movement speed penalty increased
- tweak: Egui movement speed penalty decreased
- tweak: Xingtian movement speed penalty adjusted and decreased