Loadout Additions (#680)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Atmos Techs and the Chief Engineer can get the Handheld Atmos Monitor |
all of Engineering can get the Handheld Power Monitor | Chaplain can get
a box of infinite candles | HoS can get the Terminus (Oni players
rejoice, you have a disabler)

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- Done

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![1](https://github.com/user-attachments/assets/d389a856-9977-41c5-b949-943e8fa4bb07)

![2](https://github.com/user-attachments/assets/d3c7519b-b011-4e70-bd3d-8619919aac08)

![3](https://github.com/user-attachments/assets/c35802c9-a1a6-4d85-8b88-250833f44ddf)

![4](https://github.com/user-attachments/assets/c9e4486a-8a39-4604-9bb9-85900582d463)

![5](https://github.com/user-attachments/assets/7401f41d-d3b7-42bb-8f4e-c48f782db4b8)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Handheld Atmos Monitor  >  Atmos Tech + Chief Engineer loadouts
- add: Handheld Power Monitor  >  All Engineering loadouts
- add: Infinite Candle Box  >  Chaplain loadout
- add: Antique Terminus  >  Head of Security loadout