Add Pronouns to Crew Manifest (#705)
---------

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds pronouns to the crew manifest, ported from Floofs port of DeltaVs
original PR. Original Floof PR:
https://github.com/Fansana/floofstation1/pull/630

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Cherry-pick the PR (editors note: WHY THE FUCK DID IT GO SO
SMOOTHLY)

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/ff974a58-16da-4a49-8abb-33fc468bbf90)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Pronouns in bio (blue hair not included)

Co-authored-by: Forzii <github@pawsitive.mozmail.com>
Co-authored-by: Joe <joe@faintspeaker.com>
Co-authored-by: TGRCDev <tgrc@tgrc.dev>