Barrier Patch 2: Tokyo Drift (#730)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

- Reinforces prober
- Minor prayer room alterations
- Makes maints more interesting
- Adds BEES
- Adds more maints loot
- Adds more STUFF in general
- Adds nonlethals to armory
- Powers unpowered scrubber

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Fox
- add: Added the Illegal Gorilla Room to Barrier
- tweak: Made Barrier 8% more interesting and 4% safer