Add some Funky Filk Woooooo (Let's Try This Again...) (#729)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds some jukebox music!! The only possible downside is that I/we do not
have explicit permission to use this music. However, the label that owns
the songs is defunct, and the songs are 40 years old, so I doubt any of
the science fiction lovers that wrote and sung these songs would have
any qualms with including them in this game. Of course, if any of 'em
care enough to ask us to remove them... we can absolutely do so!

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->
nothin :3

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->


# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added filk songs to the jukebox :cat:

Co-authored-by: Azzy <azzydev@vern.cc>