Feat: Basic Dynamic Power Draw for IPCs Based on Movement Speed (#696)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Made the power draw on IPCs relative to their velocity and if they are
actively inputting commands.
I.e. standing still would be a 90% reduction of base power draw, and
full sprint would be no reduction.

Based off
[this](https://discord.com/channels/1301753657024319488/1378818845438902522)
discussion

I've tested dragging someone with moon boots in a moving shuttle to make
sure it wasn't drawing more than it was supposed to.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

N/A

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: IPCs now use significantly less power when standing still