Port Crossbows From Frontier (#718)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Ports Crossbows from Frontier Station
https://github.com/new-frontiers-14/frontier-station-14/pull/1971, The
Modern Crossbow variant is available in the security techfabs once
researched through arsenal research, as well as uranium and incendiary
bolts in their respective arsenal categories. The Improvised Crossbow as
well as glass shard bolts and wooden stakes are craftable in the
construction menu.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added Crossbows! The Modern variant is available in arsenal
research and can be fabricated at the security techfab once researched
along with its bolts, The Improvised variant is craftable in the
construction menus as well as improvised glass shard bolts and wooden
stakes.