Mrrp, Mew, Mrreow (#748)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Deletes the surplus crates from the LPO/Pirate Radio

---

:cl: Fox
- remove: Removed Pirate Radio's surplus crates.