Fox Hits Me when Nobodies Watching ;( (#761)
Oh hallo! Guten abend! Ich bin eine fuckhead. Fox hates when I butcher
such a wonderful language! So I'll do it some more. Anyways, tonight on
Gregg Gear, Gregg receives a first-hand account of Birmingham's Crime
Culture, Mono stalks a Moose for several Kilometers in a hunger induced
haze and TJ recovers from the multi-year coma that being a Project
Manager of Delta-V caused.

Morgan achieves Nirvana alongside the spirit of Mr. Cobain, KC
accidentally fulfills an ancient Mesopotamian prophecy by eating four
hundred and twenty edibles and Jakumba finds himself in a mighty pickle
as he's picked apart by horny moth women.

Another stunning episode is in store for you tonight, live on Gregg Gear
_Live Internet Production: Gregg Gear, is not responsible for any
injuries, insanities or otherwise institutionalizations that may occur
from reading about, listening to or attempting to replicate any events
portrayed._



![rallytothequeen](https://github.com/user-attachments/assets/8c1018fd-a4e6-4f70-b5d4-73b52ae549e1)


<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

**BAIKAL:**
Removed cameras from the Mysta, CE and Psych's OFFICE, removed cameras
from the Chaplain, NT Rep and Mantis's BEDROOMS.
There are cameras OUTSIDE of these areas still, but all of these rooms
have curtains/shutters to prevent people from peeking in when the
inhabitants do not want to be peeped on.
(Added shutters to the Warden's Bedroom)
Also removed the weird double sided wall in the dorms area, now that
Subtle and what not cannot be heard through walls.

**TORTUGA:**
Attempted to lazily revamp the Chapel into being a multi-denominational
place of worship, and added a Sophia to the 'outdoor' area the Oracle is
at.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Re-Grow my hair
- [x] Stave off the next Mapping Quota (Fox is strict and hits me with a
water balloon filled with hot sauce if I miss it)

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![caption](https://github.com/user-attachments/assets/fd8f8dff-8dab-4ebd-b029-b7a936984651)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added Sophia to Tortuga, Glimmer may be more easily monitored.
- tweak: Tweaked the Tortuga Chapel, attempting to make it
multi-denominational.
- remove: Removed cameras from certain Offices and Bedrooms on Baikal