Add Missing Part of "Vulp Points Hands" Sprite (#767)
# Description

I'm always trying to figure out ways to make money. Money's a pretty big
thing in our society. Gotta have it. But I've had trouble making it. For
example, the idea I worked on pretty hard, was uh, this picture window
thing. I figured if I can get myself some salesmen, and go out in the
suburbs—where they have lots of picture windows—we could really make a
killing.

See, we were gonna sell... two windows, and inside would be some sand,
and ants. And then you'd have a, uh, a picture window ant colony. And
then part of the thing would be a sign, in neon, outside flashing off
and on at night, saying, oh... "Don't bug me," or, uh, "Why aren't you
working?", or uh, "It's better than television," you know, little
sayings, uh, that we could have. But I couldn't get any salesmen
interested and... it fell apart.

So I figured the- the next idea that I have, is gonna go because it's
much simpler. It's uh, a bury-it-yourself time capsule. See, I would
sell this in drug stores and uh, places like that—not in airports, but,
mm, places where they'd be interested. And it'd have a scroll inside of
it, with blank spots. It would say, "My name was"—and then _blank_—"And
I lived at"—and then _blank_—"And I used to do"—_blank_—"for a living.
And I used to believe in"—and then _blank_. And we'd have this thing,
and they—people would fill it in, and it would sell for 98 cents, and
then they'd go out in their backyard or out in the meadow and dig a
hole, and bury it.

And then, two thousand years from now—with soil erosion—all these little
time capsules would be sticking their little heads up, and Margaret
Mead's, uh, progeny would be coming around, picking 'em up, opening them
and they'd know all about us! Like, uh... we might have something? Why
should only banks, and great big buildings have time capsules?

Now, if you're interested... get in touch with me, okay?

---

## Media


![image](https://github.com/user-attachments/assets/50e464f7-cc75-4386-9e27-ed07ae6692fd)

---

# Changelog

:cl:
- fix: Fixed "Vulpkanin Points Hands" side view sprite.