Some Clothes! (#645) (#755)
**Cherrypick from Floof PR Some Clothes - Will Update Name when Ready!
(#645)**

* Rodentia Ears to Felinids per thread request

* Vulpkanin ears and tails to Rodentia

* comma my beloathed

* Flamenco Dress

Ported from Coyote Bayou.

A frilly little number for you to wear when you spin on peoples heads
after pulping them. Adds it to loadout and the theater vending machine.

* Some Clothes

* Banded dancers garb

* Red Sideless Dress

* Purple Nightclub Dress

* Black Petticoat Dress

* Blue Fancy Lad Outfit

* Cleans up capital letters and an oopsie

* That's a big L for me chief.

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Description.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added fun :D
- tweak: Tweaked fun
- fix: Fixed fun!
- remove: Removed fun :(