Morgan Gives Me Ghost Backshots when We're AGhosted (#788)
Hello hello helloooo~! Morgan yelled at me to "Drop my pants and touch
my toes, she's gonna show me where the wild kibby goes!". I'm not sure
what she means by this, but I believe in the the heart of the cards.
It's not fate, it's destiny *wild nickleback-esque guitar riffs*.

Anyways, tonight on Gregg Gear: Gregg gets his head stuck in a fence and
needs to be fully soaked in butter to be wriggled free, Fox duels a
Embden Goose over the usage of the local bus systems and KC engages in a
pie eating contest against ghost of that weaselly traitor fuck Stonewall
Jackson.

Mono receives other worldly visions after having a temperature of 120°F
(it's okay, Mono and the Moose from last episode have entered an uneasy
alliance and the Moose is caring for him), TJ goes on an epic journey to
save Valiant from the clutches of the Taliban and June is stranded on
the Moon with nothing but her big, bug ass to keep her company. (She can
breathe because she's just that cool).

Jenn flys to Vietnam and experiences a culinary master-tour of such a
wonderful land, April gets trapped in VR matrix-style and has to beat
the final boss of VR, Gabe Newell. Finally, Cami gets tied to a chair
and told how wonderful and important she is to everyone.


![realasitgets](https://github.com/user-attachments/assets/346a8628-78e1-4c95-8082-9a9510ae372d)


<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

In all seriousness, we have two map updates today.

**TORTUGA:**
Departmental Comms Console
Office Chairs galore!!!
AI Job Slot
AI Turrets

**ASTERISK:**
AI Job Slot
AI Turrets
Moved a single Wall Light off a window in the conference room
Departmental Comms Console

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Pay June Child Support
- [x] Fix the things Fox whipped me into fixing (oof ow ouchie!

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![aca8ef69-b2db-47e1-883a-a01bf8e1f39c](https://github.com/user-attachments/assets/a1ba11c7-3692-49f4-9956-ed5473b1414f)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added AI Job Slots and AI Turrets to Tortuga and Asterisk

---------

Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>