Various Kitsune Fixes (#792)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Merge of a handful of Kitsune fixes from Floof (plus another that was
also a part of kitsune fixes), all linked below:
- https://github.com/Fansana/floofstation1/pull/981
- https://github.com/Fansana/floofstation1/pull/1006
- https://github.com/Fansana/floofstation1/pull/1030

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Various Floof Contributions
- fix: Cloning should now properly replicate the original body's
appearance.
- fix: Female kitsune should now no longer laugh like a man when in fox
form.
- fix: Kitsune foxfire will now emit light, regardless of how dark the
color is.
- fix: Kitsune should now no longer weigh as much as a feather.
- fix: Kitsune can now be healed by topicals.
- fix: Kitsune can no longer place infinite foxfires.
- fix: You will now see the proper message when a kitsune shifts into
fox form.

---------

Co-authored-by: M3739 <47579354+M3739@users.noreply.github.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>