Skivvies Update! (#751)
**Based on Floof PR (#801)**

* Adds a thing that lets you take off your bra, or your panties. Also
lets people put on 2 pairs of undies.

* hey stop committing random shit

* Adds floof comments, reverts the hair thing I couldnt get working

* Fixes the thing not compiling, cutely

* Add a missing comment

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Description.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>
You can wear up to two of underwear and undershirts, and have the
ability to remove each one separately.


![image](https://github.com/user-attachments/assets/ab59e56c-1e84-40cf-9ba0-204e9cce16cc)


![image](https://github.com/user-attachments/assets/d278d8b8-9f4a-470a-ad24-25f2d41d5c9a)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: added underwear and the ability to remove it

---------

Co-authored-by: Superlagg <sprambersonson@gmail.com>