Ports Randomized Candy From Delta V (Or At Least Attempt To) (#791)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Ports PR #1074 from Delta V, adds randomized colors and flavors to the
gumballs and lollipops dispensed by medical borgs.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Get someone to test if it actually works because it's old as fuck
code.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Gumballs and lollipops dispensed by medical borgs now come in
different flavors.

---------

Co-authored-by: no <165581243+pissdemon@users.noreply.github.com>