Ough the Poster Oughhhhhhhhh on My Syndie Shiiiiiiiiiiiittttttttttt (#787)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

New LPO map with 3 spawns, a bunch of new goodies, and chief of all
FISHING!!!!
---

:cl:
- add: Added a new and improved listening post to replace the old one!