This Branch Was Originally a Markings Branch (#806)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Added a bunch of existing markings to Kitsune, added yip and yap (new
emote) to Kitsune and Vulp, other various things
edit as i forgot to link by mistake:
https://github.com/Fansana/floofstation1/pull/439 (not properly
cherry-picked due to github shenanigans)

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added the Yap emote from Floof (credit to Memeji)
- new: Added Kagebun to racial languages trait group for non-kitsune
- tweak: Added kitsune to the whitelist for a bunch of markings
- fix: Gave Kitsune yip emote (as originally intended)
- tweak: Removed restriction on some hairs