Denu! (#810)
# Description

It's like a menu... but for the Den... Denu! Or Jennu, if you're Jenn. I
dunno.

Everything in this is handled on the client-side atm. I really don't
want to fuck about trying to run this on the serverside, which means the
way I parse the messages is a little questionable, but if you're
optimistic, it's actually a feature, as it now gets applied to commands
too. Could be useful for some things.



![image](https://github.com/user-attachments/assets/71cb7995-9017-43de-b1a9-b9973c3f466c)


![image](https://github.com/user-attachments/assets/70a3962a-4e56-493f-9c3d-220424f9c264)


# Changelog

:cl:
- add: The Denu! A new menu located next to the chat filter button with
various buttons, including a typing toggle, automatic chat formatting,
and color sliders for said formatting.

---------

Co-authored-by: Cam <Nop>