Small Tweaks and Fixes (#805)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Krav Maga's Leg Sweep adjusted to only do knockdown instead of
paralysis, as well as a 5 second use delay on it to prevent potential
looping, Fixes text for the paradox anomaly examine text to be more
grammatically correct, Fixes IPCS not being able to set poses or take
off underwear, Makes it so Diamonds should only be able to be inserted
in the respective lathes that can use them, Attack speed fixes for
katanas and claymores after an old PR fucked them up, Stamina cost for
greatsword lowered to 20 from 30.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Tweaked Krav Maga's leg sweep to have a 5 second use delay
instead of 3 and to do a knockdown instead of a complete paralysis.
- tweak: Diamonds should now only be able to be inserted into the
respective lathes that use them, security techfab and exosuit
fabricator.
- tweak: Stamina cost for heavy swings on Greatsword lowered from 30 to
20.
- fix: Fixed the heavy swing speed for katanas and claymores to be more
sensible.
- fix: Fixed IPCs not being able to set poses.