Lower Salvage Magnet/Expeditions Restrictions (#824)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Done as a
[bounty](https://discord.com/channels/1301753657024319488/1384321719522558023)
>w>

Personal recommendation: make the expeditions console a restricted item
(to salvage/logistics) so Mr john scientist antag doesnt just be boring
and get away with spamming expeds on their 3x3 shittle

Salvage magnet requires the list to refresh on the station before it
will properly display but, it works!

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Expeditions work on player built shuttles
- [x] Salvage Magnet works on player built shuttles

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Expeditions can be performed on player built shuttles
- tweak: Salvage magnet works on player built shuttles