New Lobby Songs (#815)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

# LICENSE: MIT

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->
This PR is licensed as MIT because I am not the first guy that has typed
out song attributions ever. Anyone can do this. Anyone can use these
songs.
18 deletions, 8-9 song adds, and the unceremonious nuking of the
Parkstation tracks because they weren't being used.
This PR changes up the songs in the lobby.
It also fixes song attributions, which were _disorganized, missing, and
**really fucking bad, since some of the songs with aforementioned
missing credits were under CC BY and VIOLATING THEIR LICENSE**!!!_
If there's any metadata that I'm missing that is going to cause things
to show up as "Unknown song by unknown artist," let me know and I'll
fill that in.
I will gladly take any suggestions for music that we have the license to
use. I will also be sending the majority of tracks used in this PR to
maintainers and any contributor who wants to hear them.

Songs added (they are all CC licensed, and safe to use:)
- A cover of Shibamata I made, entitled "Battle! Shibamata," it's in the
Pokemon Black and White soundfont. This PR is its source.
- The Hyperthetical from Vast Error Volume 5 Side 2
- Untold Horrors from Duodenary (also from Vast Error's bandcamp)
- Cold Water, a vaporwave track by an artist named TVAAAS
- Drone ConscioUsness, by the aforementioned TVAAAS and someone named
Suri
- Shelter Dog from the Adventures with Anxiety soundtrack.
- Space Asshole (iconic.)
- Spac Stac (Song I associate most with Wiz.)
- Antarctic Reinforcement (Structured Mix) - This one is the demo
version from the Calamity Demos album. It's in weird copyright territory
given the version of it on the full Calamity album is labeled All
Rights, but Dokuro himself did say credit is paramount when it comes to
using his songs, so, oh well.
- Yoshi's Island by Glass Beach, juni's favorite song from that album.
Hi juni ❤️

Songs removed:
- Undefined Theory
- DOS = HIGH, UMB
- Comet Halley
- Hackers
- Marhaba
- Flip Flap
- Nights of Orion
- Superposition
- Endless Space
- Every Light Is Blinking At Once
- A Different Reality (Lagoona remix)
- Honestly just look at the files deleted in the diff, really. It's a
lot.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Clean out most of the unused tracks
- [x] Fix attributions
- [X] Try not to bloat the PR with songs from Vast Error because that is
my biggest source of CC-licensed songs at the moment
- [ ] Add 11 more CC licensed songs that we have permission to use

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added a lot of new music to a lobby near you.
- remove: Removed a lot of music from a lobby near you.

---------

Co-authored-by: TheDen-Bot <thelargetrucksyndicate@gmail.com>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>